requirements.txt

Flask
textblob
nltk
pyspellchecker
Werkzeug
